python/pyproject.toml
[tool.poetry]
name = "devops-scripts"
version = "0.1.0"
description = "Various scripts for BanklessDAO DevOps"
authors = ["BankelssDAO DevOps <noreply@bankless.community>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
PyGithub = "^1.55"
requests = "^2.27.1"
argparse = "^1.4.0"
pylint = "^2.12.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
